src/main/java/com/example/emantrana/DSA/String/KMP.java
package com.example.emantrana.DSA.String;

public class KMP {

    private String pattern;
    private int[] lps;

    public KMP(String pattern) {
        this.pattern = pattern;
        computeLPS();
    }

    private void computeLPS() {
        int m = pattern.length();
        lps = new int[m];
        int len = 0;
        int i = 1;

        while (i < m) {
            if (pattern.charAt(i) == pattern.charAt(len)) {
                len++;
                lps[i] = len;
                i++;
            } else {
                if (len != 0) {
                    len = lps[len - 1];
                } else {
                    lps[i] = 0;
                    i++;
                }
            }
        }
    }

    public void search(String text) {
        int n = text.length();
        int m = pattern.length();
        int i = 0; // Index for text
        int j = 0; // Index for pattern

        while (i < n) {
            if (pattern.charAt(j) == text.charAt(i)) {
                i++;
                j++;
            }

            if (j == m) {
                System.out.println("Pattern found at index " + (i - j));
                j = lps[j - 1];
            } else if (i < n && pattern.charAt(j) != text.charAt(i)) {
                if (j != 0) {
                    j = lps[j - 1];
                } else {
                    i++;
                }
            }
        }
    }

}
